Implementando outras exceções

+++ dscommerce/src/main/java/io/github/marrafon91/dscommerce/services/ProductService.java
@@ -3,11 +3,15 @@ package io.github.marrafon91.dscommerce.services;
 import io.github.marrafon91.dscommerce.dto.ProductDTO;
 import io.github.marrafon91.dscommerce.entities.Product;
 import io.github.marrafon91.dscommerce.repositories.ProductRepository;
+import io.github.marrafon91.dscommerce.services.excptions.DatabaseExecption;
 import io.github.marrafon91.dscommerce.services.excptions.ResourceNotFoundExecption;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 @Service
@@ -39,15 +43,26 @@ public class ProductService {
 
     @Transactional
     public ProductDTO update(Long id, ProductDTO dto) {
+       try {
         Product entity = repository.getReferenceById(id);
         copyDtoToEntity(dto, entity);
         entity = repository.save(entity);
         return new ProductDTO(entity);
+       } catch (EntityNotFoundException e) {
+            throw new ResourceNotFoundExecption("Prdouto não encontrado");
+       }
     }
 
-    @Transactional
+    @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id) {
+        if (!repository.existsById(id)) {
+            throw new ResourceNotFoundExecption("Recurso não encontrado");
+        }
+        try {
        repository.deleteById(id);
+        } catch (DataIntegrityViolationException e) {
+            throw  new DatabaseExecption("Falha de integridade referencial");
+        }
     }
 
     private void copyDtoToEntity(ProductDTO dto, Product entity) {
